# modules/headless-PhoneBook/headless-PhoneBook-impl/rest-openapi.yaml
components:
    schemas:
...
        Department:
            description:
                Contains all of the data for a single department.
            properties:
                id:
                    description:
                        The department internal ID.
                    format: int64
                    type: integer
                name:
                    description:
                        The department name.
                    type: string
            type: object
...
        Person:
            description:
                Contains all of the data for a single person.
            properties:
                department:
                    $ref: "#/components/schemas/Department"
                email:
                    description:
                        The person email.
                    type: string
                faxNumber:
                    description:
                        The person fax number.
                    type: string
                firstName:
                    description:
                        The person first name.
                    type: string
                id:
                    description:
                        The person internal ID.
                    format: int64
                    type: integer
                lastName:
                    description:
                        The person last name.
                    type: string
                localPhoneNumber:
                    description:
                        The person local phone number.
                    type: string
                phoneNumber:
                    description:
                        The person phone number.
                    type: string
                role:
                    $ref: "#/components/schemas/Role"
                roomNumber:
                    description:
                        The person room number.
                    type: string
                website:
                    description:
                        The person website.
                    type: string
        Role:
            description:
                Contains all of the data for a single role.
            properties:
                id:
                    description:
                        The role internal ID.
                    format: int64
                    type: integer
                name:
                    description:
                        The role name.
                    type: string
            # department:
            # $ref: "#/components/schemas/Department"
            type: object
info:
    description:
        "HeadlessPhoneBook REST API"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "HeadlessPhoneBook"
    version: v1.0
openapi: 3.0.1
paths:
...
    "/persons/{personId}":
        get:
            description:
                Retrieves the person via its ID.
            operationId: getPerson
            parameters:
                - in: path
                  name: personId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    description:
                        desc
            tags: ["Person"]
        patch:
            description:
                Replaces the person with the information sent in the request body. Any missing fields are deleted,
                unless they are required.
            operationId: patchPerson
            parameters:
                - in: path
                  name: personId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: departmentId
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: roleId
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Person"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Person"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    description:
                        desc
            tags: ["Person"]
        put:
            description:
                Replaces the person with the information sent in the request body. Any missing fields are deleted,
                unless they are required.
            operationId: putPersonApi
            parameters:
                - in: path
                  name: personId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: departmentId
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: roleId
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Person"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Person"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    description:
                        Default Response
            tags: ["Person"]
...
    "/sites/{siteId}/excel":
        get:
            description:
                Retrieves all Persons to excel
            operationId: getPersonsExcel
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: filter
                  required: false
                  schema:
                      type: string
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/octet-stream:
                            schema:
                                format: binary
                                type: string
                    description:
                        ""
            tags: ["Person"]
        post:
            description:
                Create Person from excel
            operationId: postPersonExcel
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    multipart/form-data:
                        schema:
                            properties:
                                file:
                                    format: binary
                                    type: string
                            type: object
            responses:
                200:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        "excel file successfully imported"
            tags: ["Person"]
...
    "/sites/{siteId}/persons":
        get:
            description:
                Retrieves the list of persons. Results can be paginated, filtered, searched, and sorted.
            operationId: getPersonsPage
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: filter
                  schema:
                      type: string
                - in: query
                  name: page
                  schema:
                      type: integer
                - in: query
                  name: pageSize
                  schema:
                      type: integer
                - in: query
                  name: search
                  schema:
                      type: string
                - in: query
                  name: sort
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Person"
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: "#/components/schemas/Person"
                                type: array
                    description:
                        desc
            tags: ["Person"]
        post:
            description:
                Create a new person.
            operationId: postPerson
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: departmentId
                  schema:
                      format: int64
                      type: integer
                - in: query
                  name: roleId
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/Person"
                    application/xml:
                        schema:
                            $ref: "#/components/schemas/Person"
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    description:
                        desc
            tags: ["Person"]
        put:
            description:
                Deletes persons.
            operationId: deletePersons
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            items:
                                format: int64
                                type: integer
                            type: array
                    application/xml:
                        schema:
                            items:
                                format: int64
                                type: integer
                            type: array
            responses:
                204:
                    content:
                        application/json: {}
                        application/xml: {}
                    description:
                        "Person successfully delete"
            tags: ["Person"]
    "/sites/{siteId}/persons/{personId}":
        delete:
            description:
                Deletes the person and returns a 204 if the operation succeeds.
            operationId: deletePersonApi
            parameters:
                - in: path
                  name: siteId
                  required: true
                  schema:
                      format: int64
                      type: integer
                - in: path
                  name: personId
                  required: true
                  schema:
                      format: int64
                      type: integer
            responses:
                204:
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Person"
                        application/xml:
                            schema:
                                $ref: "#/components/schemas/Person"
                    description:
                        "Person successfully delete"
            tags: ["Person"]
...